ModelingToolkitStandardLibrary: Magnetic Components

ModelingToolkitStandardLibrary: Magnetic Components

Index

Flux Tubes

Index

ModelingToolkitStandardLibrary.Magnetic.FluxTubes.ConstantMagneticFlux

ModelingToolkitStandardLibrary.Magnetic.FluxTubes.ConstantMagneticPotentialDifference

ModelingToolkitStandardLibrary.Magnetic.FluxTubes.ConstantPermeance

ModelingToolkitStandardLibrary.Magnetic.FluxTubes.ConstantReluctance

ModelingToolkitStandardLibrary.Magnetic.FluxTubes.Crossing

ModelingToolkitStandardLibrary.Magnetic.FluxTubes.EddyCurrent

ModelingToolkitStandardLibrary.Magnetic.FluxTubes.ElectroMagneticConverter

ModelingToolkitStandardLibrary.Magnetic.FluxTubes.Ground

ModelingToolkitStandardLibrary.Magnetic.FluxTubes.Idle

ModelingToolkitStandardLibrary.Magnetic.FluxTubes.NegativeMagneticPort

ModelingToolkitStandardLibrary.Magnetic.FluxTubes.PositiveMagneticPort

ModelingToolkitStandardLibrary.Magnetic.FluxTubes.Short

ModelingToolkitStandardLibrary.Magnetic.FluxTubes.TwoPort

Flux Tubes

Flux Tube Utilities

ModelingToolkitStandardLibrary.Magnetic.FluxTubes.PositiveMagneticPort — Constant

Positive magnetic port

ModelingToolkitStandardLibrary.Magnetic.FluxTubes.NegativeMagneticPort — Constant

Negative magnetic port

ModelingToolkitStandardLibrary.Magnetic.FluxTubes.TwoPort — Constant

TwoPort(; name, V_m = 0.0, Phi = 0.0)

Partial component with magnetic potential difference between two magnetic ports p and n and magnetic flux Phi from p to n.

Parameters:

V_m: Initial magnetic potential difference between both ports

Phi: Initial magnetic flux from portp to portn

Basic Flux Tube Blocks

ModelingToolkitStandardLibrary.Magnetic.FluxTubes.Ground — Constant

Ground(; name)

Zero magnetic potential.

ModelingToolkitStandardLibrary.Magnetic.FluxTubes.Idle — Constant

Idle(;name)

Idle running branch.

ModelingToolkitStandardLibrary.Magnetic.FluxTubes.Short — Constant

Short(;name)

Short cut branch.

ModelingToolkitStandardLibrary.Magnetic.FluxTubes.Crossing — Constant

Crossing(;name)

Crossing of two branches.

This is a simple crossing of two branches. The ports portp1 and portp2 are connected, as well as portn1 and portn2.

ModelingToolkitStandardLibrary.Magnetic.FluxTubes.ConstantPermeance — Constant

ConstantPermeance(; name, G_m = 1.0)

Constant permeance.

Parameters:

G_m: [H] Magnetic permeance

ModelingToolkitStandardLibrary.Magnetic.FluxTubes.ConstantReluctance — Constant

ConstantReluctance(; name, R_m = 1.0)

Constant reluctance.

Parameters:

R_m: [H^-1] Magnetic reluctance

ModelingToolkitStandardLibrary.Magnetic.FluxTubes.EddyCurrent — Constant

EddyCurrent(;name, Phi, rho = 0.098e-6, l = 1, A = 1)

For modelling of eddy current in a conductive magnetic flux tube. Initial magnetic flux flowing into the port_p can be set with Phi ([Wb])

Parameters:

rho: [ohm * m] Resistivity of flux tube material (default: Iron at 20degC)

l: [m] Average length of eddy current path

A: [m^2] Cross sectional area of eddy current path

ModelingToolkitStandardLibrary.Magnetic.FluxTubes.ElectroMagneticConverter — Constant

ElectroMagneticConverter(; name, N, Phi)

Ideal electromagnetic energy conversion.

The electromagnetic energy conversion is given by Ampere's law and Faraday's law respectively V_m = N * i N * dΦ/dt = -v

Initial magnetic flux flowing into the port_p can be set with Phi ([Wb])

Parameters:

N: Number of turns

Flux Tube Sources

ModelingToolkitStandardLibrary.Magnetic.FluxTubes.ConstantMagneticPotentialDifference — Constant

ConstantMagneticPotentialDifference(; name, V_m = 0.0)

Constant magnetomotive force.

Parameters:

V_m: [A] Magnetic potential difference

ModelingToolkitStandardLibrary.Magnetic.FluxTubes.ConstantMagneticFlux — Constant

ConstantMagneticFlux(; name, Phi = 0.0)

Source of constant magnetic flux.

Parameters:

Phi: [Wb] Magnetic flux