ModelingToolkitStandardLibrary: Thermal Components

Thermal Utilities

ModelingToolkitStandardLibrary.Thermal.HeatPort — Function

Port for a thermal system.

Thermal Components

ModelingToolkitStandardLibrary.Thermal.BodyRadiation — Function

BodyRadiation(; name, G=1.0)

Lumped thermal element for radiation heat transfer.

Parameters:

G: [m^2] Net radiation conductance between two surfaces

ModelingToolkitStandardLibrary.Thermal.ConvectiveConductor — Function

ConvectiveConductor(; name, G=1.0)

Lumped thermal element for heat convection.

Parameters:

G: [W/K] Convective thermal conductance

ModelingToolkitStandardLibrary.Thermal.ConvectiveResistor — Function

ConvectiveResistor(; name, R=1.0)

Lumped thermal element for heat convection.

Parameters:

R: [K/W] Constant thermal resistance of material

ModelingToolkitStandardLibrary.Thermal.HeatCapacitor — Function

HeatCapacitor(; name, C=1.0, T_start=293.15 + 20)

Lumped thermal element storing heat

Parameters:

C: [J/K] Heat capacity of element (= cp*m)

T_start: Initial temperature of element

ModelingToolkitStandardLibrary.Thermal.ThermalConductor — Function

ThermalConductor(;name, G=1.0)

Lumped thermal element transporting heat without storing it.

Parameters:

G: [W/K] Constant thermal conductance of material

ModelingToolkitStandardLibrary.Thermal.ThermalResistor — Function

ThermalResistor(; name, R=1.0)

Lumped thermal element transporting heat without storing it.

Parameters:

R: [K/W] Constant thermal resistance of material

ModelingToolkitStandardLibrary.Thermal.ThermalCollector — Function

ThermalCollector(; name, m=1)

Collects m heat flows

This is a model to collect the heat flows from m heatports to one single heatport.

Parameters:

m: Number of heat ports (e.g. m=2: port_a1, port_a2)

Thermal Sensors

ModelingToolkitStandardLibrary.Thermal.RelativeTemperatureSensor — Function

RelativeTemperatureSensor(; name)

Relative Temperature sensor.

The relative temperature port_a.T - port_b.T is determined between the two ports of this component and is provided as output signal in kelvin.

ModelingToolkitStandardLibrary.Thermal.HeatFlowSensor — Function

HeatFlowSensor(; name)

Heat flow rate sensor.

This model is capable of monitoring the heat flow rate flowing through this component. The sensed value of heat flow rate is the amount that passes through this sensor while keeping the temperature drop across the sensor zero. This is an ideal model so it does not absorb any energy and it has no direct effect on the thermal response of a system it is included in. The output signal is positive, if the heat flows from port_a to port_b.

ModelingToolkitStandardLibrary.Thermal.TemperatureSensor — Function

TemperatureSensor(; name)

Absolute temperature sensor in kelvin.

This is an ideal absolute temperature sensor which returns the temperature of the connected port in kelvin as an output signal. The sensor itself has no thermal interaction with whatever it is connected to. Furthermore, no thermocouple-like lags are associated with this sensor model.

Thermal Sources

ModelingToolkitStandardLibrary.Thermal.FixedHeatFlow — Function

FixedHeatFlow(; name, Q_flow=1.0, T_ref=293.15, alpha=0.0)

Fixed heat flow boundary condition.

This model allows a specified amount of heat flow rate to be "injected" into a thermal system at a given port. The constant amount of heat flow rate Q_flow is given as a parameter. The heat flows into the component to which the component FixedHeatFlow is connected, if parameter Q_flow is positive.

Parameters:

Q_flow: [W] Fixed heat flow rate at port

T_ref: [K] Reference temperature

alpha: [1/K] Temperature coefficient of heat flow rate

ModelingToolkitStandardLibrary.Thermal.FixedTemperature — Function

FixedTemperature(; name, T=0.0)

Fixed temperature boundary condition in kelvin.

This model defines a fixed temperature T at its port in kelvin, i.e., it defines a fixed temperature as a boundary condition.

Parameters:

T: [K] Fixed temperature boundary condition

Missing docstring.

Missing docstring for ThermalGround. Check Documenter's build log for details.