ModelingToolkitStandardLibrary: Electrical Components

Electrical Utilities

ModelingToolkitStandardLibrary.Electrical.Pin — Function

Port for an electrical system.

Analog Components

ModelingToolkitStandardLibrary.Electrical.Capacitor — Function

Capacitor(;name, C=1.0, v_start=0.0)

Ideal linear electrical capacitor.

Parameters:

C: [F] Capacity

v_start: [V] Initial voltage of capacitor

ModelingToolkitStandardLibrary.Electrical.Ground — Function

Ground(;name)

Ground of an electrical circuit. The potential at the ground node is zero. Every circuit must have one ground node.

ModelingToolkitStandardLibrary.Electrical.Inductor — Function

Inductor(;name, L=1.0e-6, i_start=0.0)

Ideal linear electrical inductor.

Parameters:

L: [H] Inductance

i_start: [A] Initial current through inductor

ModelingToolkitStandardLibrary.Electrical.Resistor — Function

Resistor(;name, R=1.0)

Ideal linear electrical resistor.

Parameters:

R: [Ω] Resistance

Missing docstring.

Missing docstring for Short. Check Documenter's build log for details.

ModelingToolkitStandardLibrary.Electrical.IdealOpAmp — Function

IdealOpAmp(;name)

Ideal operational amplifier (norator-nullator pair). The ideal OpAmp is a two-port. The left port is fixed to v1=0 and i1=0 (nullator). At the right port both any voltage v2 and any current i2 are possible (norator).

Analog Sensors

ModelingToolkitStandardLibrary.Electrical.CurrentSensor — Function

CurrentSensor(; name)

Sensor to measure the current through a branch in amperes.

ModelingToolkitStandardLibrary.Electrical.PotentialSensor — Function

PotentialSensor(; name)

Sensor to measure the potential in volt.

ModelingToolkitStandardLibrary.Electrical.VoltageSensor — Function

VoltageSensor(; name)

Sensor to measure the potential difference between two pins in volt.

ModelingToolkitStandardLibrary.Electrical.PowerSensor — Function

PowerSensor(; name)

Sensor to measure the power in watt.

ModelingToolkitStandardLibrary.Electrical.MultiSensor — Function

MultiSensor(; name)

Sensor to measure current [A], voltage [V] and power [W].

Analogue Sources

ModelingToolkitStandardLibrary.Electrical.ConstantVoltage — Function

ConstantVoltage(;name, V = 1.0)

Source for constant voltage,

Parameters:

V: [V] Voltage

ModelingToolkitStandardLibrary.Electrical.SineVoltage — Function

SineVoltage(;name, offset=0.0, amplitude=1.0, frequency=1.0, start_time=0.0, phase=0.0)

Generate sine voltage.

Parameters:

frequency: [Hz] Frequency of sine wave

amplitude: [V] Amplitude of sine wave

phase: [rad] Phase of sine wave

offset: [V] Offset of output voltage

start_time: [s] Output y = offset for t < start_time

ModelingToolkitStandardLibrary.Electrical.StepVoltage — Function

StepVoltage(;name, offset=0.0, start_time=0.0, height=1.0)

Generate step voltage.

Parameters:

height: [V] Height of step

offset: [V] Offset of output voltage

start_time: [s] Output y = offset for t < start_time

ModelingToolkitStandardLibrary.Electrical.RampVoltage — Function

RampVoltage(;name, offset=0.0, start_time=0.0, duration=1.0, height=1.0)

Generate ramp voltage.

Parameters:

height: [V] Height of ramp

duration: [s] Duration of ramp (= 0.0 gives a Step)

offset: [V] Offset of output voltage

start_time: [s] Output y = offset for t < start_time

ModelingToolkitStandardLibrary.Electrical.SquareVoltage — Function

SquareVoltage(; name, offset=0.0, amplitude=1.0, frequency=1.0, start_time=0.0)

Generate square voltage.

Missing docstring.

Missing docstring for TriangularVoltage. Check Documenter's build log for details.

ModelingToolkitStandardLibrary.Electrical.CosineVoltage — Function

CosineVoltage(;name, offset=0.0, amplitude=1.0, frequency=1.0, start_time=0.0, phase=0.0)

Generate cosine voltage.

Parameters:

frequency: [Hz] Frequency of sine wave

amplitude: [V] Amplitude of sine wave

phase: [rad] Phase of sine wave

offset: [V] Offset of output voltage

start_time: [s] Output y = offset for t < start_time

ModelingToolkitStandardLibrary.Electrical.ExpSineVoltage — Function

Generate damped sine voltage.

Parameters:

frequency: [Hz] Frequency of sine wave

amplitude: [V] Amplitude of sine wave

damping: [1/s] Damping coefficient of sine wave

phase: [rad] Phase of sine wave

offset: [V] Offset of output voltage

start_time: [s] Output y = offset for t < start_time

ModelingToolkitStandardLibrary.Electrical.ConstantCurrent — Function

ConstantCurrent(;name, I = 1.0)

Source for constant current.

Parameters:

I: [A] Current

ModelingToolkitStandardLibrary.Electrical.SineCurrent — Function

SineCurrent(;name, offset=0.0, amplitude=1.0, frequency=1.0, start_time=0.0, phase=0.0)

Generate sine current.

Parameters:

frequency: [Hz] Frequency of sine wave

amplitude: [A] Amplitude of sine wave

phase: [rad] Phase of sine wave

offset: [A] Offset of output current

start_time: [s] Output y = offset for t < start_time

ModelingToolkitStandardLibrary.Electrical.StepCurrent — Function

StepCurrent(;name, offset=0.0, start_time=0.0, height=1.0)

Generate step current.

Parameters:

height: [A] Height of step

offset: [A] Offset of output current

start_time: [s] Output y = offset for t < start_time

ModelingToolkitStandardLibrary.Electrical.RampCurrent — Function

RampCurrent(;name, offset=0.0, start_time=0.0, duration=1.0, height=1.0)

Generate ramp current.

Parameters:

height: [A] Height of ramp

duration: [s] Duration of ramp (= 0.0 gives a Step)

offset: [A] Offset of output current

start_time: [s] Output y = offset for t < start_time

Missing docstring.

Missing docstring for SquareCurrent. Check Documenter's build log for details.

Missing docstring.

Missing docstring for TriangularCurrent. Check Documenter's build log for details.

ModelingToolkitStandardLibrary.Electrical.CosineCurrent — Function

CosineCurrent(;name, offset=0.0, amplitude=1.0, frequency=1.0, start_time=0.0, phase=0.0)

Generate cosine current.

Parameters:

frequency: [Hz] Frequency of sine wave

amplitude: [A] Amplitude of sine wave

phase: [rad] Phase of sine wave

offset: [A] Offset of output current

start_time: [s] Output y = offset for t < start_time

ModelingToolkitStandardLibrary.Electrical.ExpSineCurrent — Function

ExpSineCurrent(;name, offset=0.0, amplitude=1.0, frequency=1.0, start_time=0.0, phase=0.0, damping=0.0)

Generate damped sine current.

Parameters:

frequency: [Hz] Frequency of sine wave

amplitude: [A] Amplitude of sine wave

damping: [1/s] Damping coefficient of sine wave

phase: [rad] Phase of sine wave

offset: [A] Offset of output current

start_time: [s] Output y = offset for t < start_time