ModelingToolkitStandardLibrary: Electrical Components

Electrical Utilities

ModelingToolkitStandardLibrary.Electrical.Pin — Function

@connector function Pin(; name)

A pin in an analog circuit.

States

v(t): [V] The voltage at this pin

i(t): [A] The current passing through this pin

ModelingToolkitStandardLibrary.Electrical.OnePort — Function

    OnePort(; name, v_start=0.0, i_start=0.0)

Component with two electrical pins p and n and current i from p to n.

Parameters:

v_start: [V] Initial voltage across the component

i_start: [A] Initial current through the component

ModelingToolkitStandardLibrary.Electrical.DigitalPin — Function

@connector function DigitalPin(; name)

A pin in a digital circuit.

States

v(t): [V] The voltage at this pin

i(t): [A] The current passing through this pin

val(t): The binary value of the pin at this point. A voltage from 0V to 0.8V is a binary value of 0. A voltage in the range 2.0V to 5.0V is 1. Any other value is X.

Analog Components

ModelingToolkitStandardLibrary.Electrical.Ground — Function

function Ground(; name)

Ground node with the potential of zero and connector g. Every circuit must have one ground node.

Connectors

g

ModelingToolkitStandardLibrary.Electrical.Resistor — Function

function Resistor(; name, R)

Creates an ideal Resistor following Ohm's Law.

States

v(t): [V] The voltage across the resistor, given by p.i * R

Connectors

p Positive pin

n Negative pin

Parameters:

R: [Ω] Resistance

ModelingToolkitStandardLibrary.Electrical.Conductor — Function

Conductor(;name, G)

Ideal linear electrical conductor.

States

see OnePort

Connectors

p Positive pin

n Negative pin

Parameters:

G: [S] Conductance

ModelingToolkitStandardLibrary.Electrical.Capacitor — Function

function Capacitor(; name, C)

Creates an ideal Capacitor.

States

v(t): [V] The voltage across the capacitor, given by D(v) ~ p.i / C

Connectors

p Positive pin

n Negative pin

Parameters:

C: [F] Capacitance

v_start: [V] Initial voltage of capacitor

ModelingToolkitStandardLibrary.Electrical.Inductor — Function

function Inductor(; name, L)

Creates an ideal Inductor.

States

v(t): [V] The voltage across the inductor, given by D(p.i) ~ v / L

Connectors

p Positive pin

n Negative pin

Parameters:

L: [H] Inductance

i_start: [A] Initial current through inductor

ModelingToolkitStandardLibrary.Electrical.IdealOpAmp — Function

function IdealOpAmp(; name)

Ideal operational amplifier (norator-nullator pair). The ideal OpAmp is a two-port. The left port is fixed to v1 = 0 and i1 = 0 (nullator). At the right port both any voltage v2 and any current i2 are possible (norator).

States

v1(t): [V] Voltage of left port

v2(t): [V] Voltage of right port

i1(t): [A] Current of left port

i2(t): [A] Current of right port

Connectors

p1 Positive pin (left port)

p2 Positive pin (right port)

n1 Negative pin (left port)

n2 Negative pin (right port)

Analog Sensors

ModelingToolkitStandardLibrary.Electrical.CurrentSensor — Function

function CurrentSensor(; name)

Creates a circuit component that measures the current flowing through it. Analogous to an ideal ammeter.

States

i(t): [A] Current through the sensor

Connectors

p

Positive pin

n Negative pin

ModelingToolkitStandardLibrary.Electrical.PotentialSensor — Function

function PotentialSensor(; name)

Creates a circuit component which measures the potential at a pin.

States

phi(t): [V] The potential at this point

Connectors

p Pin at which potential is to be measured

ModelingToolkitStandardLibrary.Electrical.VoltageSensor — Function

function VoltageSensor(; name)

Creates a circuit component that measures the voltage across it. Analogous to an ideal voltmeter.

States

v(t): [V] The voltage across this component

Connectors

p Positive pin

n Negative pin

ModelingToolkitStandardLibrary.Electrical.PowerSensor — Function

function PowerSensor(; name)

Combines a VoltageSensor and a CurrentSensor to measure the power being consumed by a circuit.

States

power(t): [W] The power being consumed, given by the product of voltage and current.

Connectors

pc Corresponds to the p pin of the CurrentSensor

nc Corresponds to the n pin of the CurrentSensor

pv Corresponds to the p pin of the VoltageSensor

nv Corresponds to the n pin of the VoltageSensor

ModelingToolkitStandardLibrary.Electrical.MultiSensor — Function

function MultiSensor(; name)

Combines a VoltageSensor and a CurrentSensor.

States

v(t): [V] The voltage across the VoltageSensor

i(t): [A] The current across the CurrentSensor

Connectors

pc Corresponds to the p pin of the CurrentSensor

nc Corresponds to the n pin of the CurrentSensor

pv Corresponds to the p pin of the VoltageSensor

nv Corresponds to the n pin of the VoltageSensor

Analog Sources

ModelingToolkitStandardLibrary.Electrical.ConstantVoltage — Function

function ConstantVoltage(; name, V)

The source for an ideal constant voltage.

States

v(t): [V] The voltage across this source, given by p.v - n.v and is always constant

Connectors

p Positive pin

n Negative pin

Parameters:

V: [V] The constant voltage across the terminals of this source

ModelingToolkitStandardLibrary.Electrical.SineVoltage — Function

function SineVoltage(; name, offset=0.0, amplitude=1.0, frequency=1.0, start_time=0.0, phase=0.0)

A source in which the voltage across its terminals is a sine function of time.

States

v(t): [V] The voltage across this source, given by p.v - n.v

Connectors

p Positive port

n Negative port

Parameters

offset: [V] A constant offset added to the voltage output

amplitude: [V] The amplitude of the sine function

frequency: [Hz] The frequency of the sine function

start_time: [s] The time at which the source starts functioning. Before this time, the voltage across its terminals is offset.

phase: [rad] The phase offset of the sine function

ModelingToolkitStandardLibrary.Electrical.StepVoltage — Function

function StepVoltage(; name, offset=0.0, start_time=0.0, height=1.0)

A source in which the voltage across its terminals increases from offset to offset+height at start_time

States

v(t): [V] The voltage across this source, given by p.v - n.v

Connectors

p Positive port

n Negative port

Observables

offset: [V] A constant offset added to the voltage output

start_time: [s] The time at which the source starts functioning, and the voltage jumps

height: [V] Magnitude of increase in voltage

ModelingToolkitStandardLibrary.Electrical.RampVoltage — Function

function RampVoltage(; name, offset=0.0, start_time=0.0, duration=1.0, height=1.0)

A source in which the voltage across grows linearly from offset to offset+height over the time interval duration starting at start_time

States

v(t): [V] The voltage across this source, given by p.v - n.v

Connectors

p Positive port

n Negative port

Parameters

offset: [V] A constant offset added to the voltage output

start_time: [s] The time at which the voltage starts growing

duration: [s] The duration of the ramp (0.0 gives a step)

height: [V] The amount that the voltage grows in the time interval

ModelingToolkitStandardLibrary.Electrical.SquareVoltage — Function

function SquareVoltage(; name, offset=0.0, amplitude=1.0, frequency=1.0, start_time=0.0)

A source in which the voltage across its terminals is a square function of time.

States

v(t): [V] The voltage across this source, given by p.v - n.v

Connectors

p Positive port

n Negative port

Parameters

offset: [V] A constant offset added to the voltage output

amplitude: [V] The amplitude of the square wave function

frequency: [Hz] The frequency of the square wave function

start_time: [s] The time at which the source starts functioning. Before this time, the voltage across its terminals is offset.

ModelingToolkitStandardLibrary.Electrical.TriangularVoltage — Function

function TriangularVoltage(; name, offset=0.0, amplitude=1.0, frequency=1.0, start_time=0.0)

A source in which the voltage across its terminals is a triangular function of time.

States

v(t): [V] The voltage across this source, given by p.v - n.v

Connectors

p Positive port

n Negative port

Observables

offset: [V] A constant offset added to the voltage output

amplitude: [V] Amplitude of the triangular wave function

frequency: [Hz] Frequency of the triangular wave function

start_time: [s] The time at which the source starts functioning. Before this, the output of the source is offset

ModelingToolkitStandardLibrary.Electrical.CosineVoltage — Function

function CosineVoltage(; name, offset=0.0, amplitude=1.0, frequency=1.0, starttime=0.0, phase=0.0)

A source in which the voltage across its terminals is a cosine function of time.

States

v(t): [V] The voltage across this source, given by p.v - n.v

Connectors

p Positive port

n Negative port

Observables

offset: [V] A constant offset added to the voltage output

amplitude: [V] The amplitude of the cosine function

frequency: [Hz] The frequency of the cosine function

starttime: [s] The time at which the source starts functioning. Before this time, the voltage across its terminals is 0.

phase: [rad] The phase offset of the cosine function

ModelingToolkitStandardLibrary.Electrical.ExpSineVoltage — Function

function ExpSineVoltage(; name, offset=0.0, amplitude=1.0, frequency=1.0, start_time=0.0, phase=0.0, damping=0.0)

A source in which the voltage across its terminals is a damped sine function of time.

States

v(t): [V] The voltage across this source, given by p.v - n.v

Connectors

p Positive port

n Negative port

Parameters

offset: [V] A constant offset added to the voltage output

amplitude: [V] The amplitude of the damped sine function

frequency: [Hz] The frequency of the damped sine function

start_time: [s] The time at which the source starts functioning. Before this time, the voltage across its terminals is offset.

phase: [rad] The phase offset of the damped sine function

damping_coef: [1/s] Damping coefficient of the damped sine function

ModelingToolkitStandardLibrary.Electrical.ConstantCurrent — Function

function ConstantCurrent(; name, I = 1.0)

The source for an ideal constant current.

States

i(t): [A] The current through this source, which is always constant

Connectors

p Positive pin

n Negative pin

Parameters:

I: [A] The constant current through the terminals of this source

ModelingToolkitStandardLibrary.Electrical.SineCurrent — Function

function SineCurrent(; name, offset=0.0, amplitude=1.0, frequency=1.0, start_time=0.0, phase=0.0)

A source in which the current through its terminals is a sine function of time.

States

i(t): [A] The current through this source

Connectors

p Positive port

n Negative port

Parameters

offset: [A] A constant offset added to the current output

amplitude: [V] The amplitude of the sine function

frequency: [Hz] The frequency of the sine function

start_time: [s] The time at which the source starts functioning. Before this time, the current through its terminals is offset.

phase: [rad] The phase offset of the sine function

ModelingToolkitStandardLibrary.Electrical.StepCurrent — Function

function StepCurrent(; name, offset=0.0, start_time=0.0, height=1.0)

A source in which the current through its terminals increases from offset to offset+height at start_time

States

i(t): [A] The current through this source

Connectors

p Positive port

n Negative port

Observables

offset: [A] A constant offset added to the current output

start_time: [s] The time at which the source starts functioning, and the current jumps

height: [A] Magnitude of increase in current

ModelingToolkitStandardLibrary.Electrical.RampCurrent — Function

function RampCurrent(; name, offset=0.0, start_time=0.0, duration=1.0, height=1.0)

A source in which the current grows linearly from offset to offset+height over the time interval duration starting at start_time

States

i(t): [A] The current through this source

Connectors

p Positive port

n Negative port

Parameters

offset: [A] A constant offset added to the current output

start_time: [s] The time at which the current starts growing

duration: [s] The duration of the ramp (0.0 gives a step)

height: [A] The amount that the current grows in the time interval

ModelingToolkitStandardLibrary.Electrical.SquareCurrent — Function

function SquareCurrent(; name, offset=0.0, amplitude=1.0, frequency=1.0, start_time=0.0)

A source in which the current through its terminals is a square function of time.

States

i(t): [A] The current through this source

Connectors

p Positive port

n Negative port

Parameters

offset: [A] A constant offset added to the current output

amplitude: [A] The amplitude of the square wave function

frequency: [Hz] The frequency of the square wave function

start_time: [s] The time at which the source starts functioning. Before this time, the current through its terminals is offset.

ModelingToolkitStandardLibrary.Electrical.TriangularCurrent — Function

function TriangularCurrent(; name, offset=0.0, amplitude=1.0, frequency=1.0, start_time=0.0)

A source in which the current through its terminals is a triangular function of time.

States

i(t): [A] The current through this source

Connectors

p Positive port

n Negative port

Observables

offset: [A] A constant offset added to the current output

amplitude: [A] Amplitude of the triangular wave function

frequency: [Hz] Frequency of the triangular wave function

start_time: [s] The time at which the source starts functioning. Before this, the output of the source is offset

ModelingToolkitStandardLibrary.Electrical.CosineCurrent — Function

function CosineCurrent(; name, offset=0.0, amplitude=1.0, frequency=1.0, starttime=0.0, phase=0.0)

A source in which the current through its terminals is a cosine function of time.

States

i(t): [A] The current through this source

Connectors

p Positive port

n Negative port

Observables

offset: [A] A constant offset added to the current output

amplitude: [A] The amplitude of the cosine function

frequency: [Hz] The frequency of the cosine function

starttime: [s] The time at which the source starts functioning. Before this time, the current through its terminals is 0.

phase: [rad] The phase offset of the cosine function

ModelingToolkitStandardLibrary.Electrical.ExpSineCurrent — Function

function ExpSineCurrent(; name, offset=0.0, amplitude=1.0, frequency=1.0, start_time=0.0, phase=0.0, damping=0.0)

A source in which the current through its terminals is a damped sine function of time.

States

i(t): [A] The current through this source

Connectors

p Positive port

n Negative port

Parameters

offset: [A] A constant offset added to the current output

amplitude: [A] The amplitude of the damped sine function

frequency: [Hz] The frequency of the damped sine function

start_time: [s] The time at which the source starts functioning. Before this time, the current through its terminals is offset.

phase: [rad] The phase offset of the damped sine function

damping_coef: [1/s] Damping coefficient of the damped sine function

Digital Gates

Missing docstring.

Missing docstring for Not. Check Documenter's build log for details.

Missing docstring.

Missing docstring for And. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Nand. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Or. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Nor. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Xor. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Xnor. Check Documenter's build log for details.

Digital Components

Missing docstring.

Missing docstring for HalfAdder. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FullAdder. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MUX. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DEMUX. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Encoder. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Decoder. Check Documenter's build log for details.

Digital Sources

Missing docstring.

Missing docstring for PulseDiff. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Set. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Reset. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Pulse. Check Documenter's build log for details.