ModelingToolkitStandardLibrary: Electrical Components

ModelingToolkitStandardLibrary: Electrical Components

Index

Electrical Utilities

Analog Components

Analog Sensors

Analog Sources

Digital Gates

Digital Components

Digital Sources

Index

ModelingToolkitStandardLibrary.Electrical.Capacitor

ModelingToolkitStandardLibrary.Electrical.Conductor

ModelingToolkitStandardLibrary.Electrical.Current

ModelingToolkitStandardLibrary.Electrical.CurrentSensor

ModelingToolkitStandardLibrary.Electrical.Diode

ModelingToolkitStandardLibrary.Electrical.Ground

ModelingToolkitStandardLibrary.Electrical.HeatingDiode

ModelingToolkitStandardLibrary.Electrical.IdealOpAmp

ModelingToolkitStandardLibrary.Electrical.Inductor

ModelingToolkitStandardLibrary.Electrical.MultiSensor

ModelingToolkitStandardLibrary.Electrical.OnePort

ModelingToolkitStandardLibrary.Electrical.Pin

ModelingToolkitStandardLibrary.Electrical.PotentialSensor

ModelingToolkitStandardLibrary.Electrical.PowerSensor

ModelingToolkitStandardLibrary.Electrical.Resistor

ModelingToolkitStandardLibrary.Electrical.VariableResistor

ModelingToolkitStandardLibrary.Electrical.Voltage

ModelingToolkitStandardLibrary.Electrical.VoltageSensor

ModelingToolkitStandardLibrary.Electrical.DigitalPin

Electrical Utilities

ModelingToolkitStandardLibrary.Electrical.Pin — Constant

Pin(; name)

A pin in an analog circuit.

States:

v(t): [V] The voltage at this pin

i(t): [A] The current passing through this pin

ModelingToolkitStandardLibrary.Electrical.OnePort — Constant

OnePort(; name, v = 0.0, i = 0.0)

Component with two electrical pins p and n and current i flows from p to n.

States:

v(t): [V] The voltage across component p.v - n.v

i(t): [A] The current passing through positive pin

Connectors:

p Positive pin

n Negative pin

ModelingToolkitStandardLibrary.Electrical.DigitalPin — Function

DigitalPin(; name)

A pin in a digital circuit.

States:

v(t): [V] The voltage at this pin

i(t): [A] The current passing through this pin

val(t): The binary value of the pin at this point. A voltage from 0V to 0.8V is a binary value of 0.

A voltage in the range 2.0V to 5.0V is 1. Any other value is X.

Analog Components

ModelingToolkitStandardLibrary.Electrical.Ground — Constant

Ground(; name)

Ground node with the potential of zero and connector g. Every circuit must have one ground node.

Connectors:

g

ModelingToolkitStandardLibrary.Electrical.Resistor — Constant

Resistor(; name, R)

Creates an ideal Resistor following Ohm's Law.

States:

See OnePort

Connectors:

p Positive pin

n Negative pin

Parameters:

R: [Ohm] Resistance

ModelingToolkitStandardLibrary.Electrical.Conductor — Constant

Conductor(; name, G)

Creates an ideal conductor.

States:

See OnePort

Connectors:

p Positive pin

n Negative pin

Parameters:

G: [S] Conductance

ModelingToolkitStandardLibrary.Electrical.Capacitor — Constant

Capacitor(; name, C, v)

Creates an ideal capacitor. Initial voltage of capacitor can be set with v ([V])

States:

See OnePort

Connectors:

p Positive pin

n Negative pin

Parameters:

C: [F] Capacitance

ModelingToolkitStandardLibrary.Electrical.Inductor — Constant

Inductor(; name, L, i)

Creates an ideal Inductor. Initial current through inductor can be set with i ([A]).

States:

See OnePort

Connectors:

p Positive pin

n Negative pin

Parameters:

L: [H] Inductance

ModelingToolkitStandardLibrary.Electrical.IdealOpAmp — Constant

IdealOpAmp(; name)

Ideal operational amplifier (norator-nullator pair). The ideal OpAmp is a two-port. The left port is fixed to v1 = 0 and i1 = 0 (nullator). At the right port both any voltage v2 and any current i2 are possible (norator).

States:

See TwoPort

Connectors:

p1 Positive pin (left port)

p2 Positive pin (right port)

n1 Negative pin (left port)

n2 Negative pin (right port)

ModelingToolkitStandardLibrary.Electrical.Diode — Constant

    Diode(; name, Is = 1e-6, n = 1, T = 300.15)

Ideal diode based on the Shockley diode equation.

States

- See [OnePort](@ref)

Connectors

- `p` Positive pin
- `n` Negative pin

Parameters

- `Is`: [`A`] Saturation current
- `n`: Ideality factor
- `T`: [K] Ambient temperature

ModelingToolkitStandardLibrary.Electrical.HeatingDiode — Constant

HeatingDiode(; name, Is = 1e-6, n = 1)

Temperature dependent diode based on the Shockley diode equation.

States

- See [OnePort](@ref)

Connectors

- `p` Positive pin
- `n` Negative pin
- `port` [HeatPort](@ref) Heat port to model the temperature dependency

Parameters:

- `Is`: [`A`] Saturation current
- `n`: Ideality factor

ModelingToolkitStandardLibrary.Electrical.VariableResistor — Constant

VariableResistor(; name, R_ref = 1.0, T_ref = 300.15, R_const = 1e-3, T_dep = false)

Variable resistor with optional temperature dependency.

The total resistance R ∈ [Rconst, Rconst + Rref], where pos is the position of the wiper and Rref is the variable resistance between p and n. The total resistance is then:

R = Rconst + pos * Rref

If T_dep is true, then R also depends on the temperature of the heat port with temperature coefficient alpha. The total resistance is then:

R = Rconst + pos * Rref * (1 + alpha * (port.T - T_ref))

States

- See [OnePort](@ref)
- `pos(t)`: Position of the wiper (normally 0-1)
- `R(t)`: Resistance

Connectors

    - `p` Positive pin
    - `n` Negative pin
    - `position` RealInput to set the position of the wiper
    - `port` [HeatPort](@ref) Heat port to model the temperature dependency

Parameters

    - `R_ref`: [`Ω`] Resistance at temperature T_ref when fully closed (pos=1.0)
    - `T_ref`: [K] Reference temperature
    - `R_const`: [`Ω`] Constant resistance between p and n
    - `T_dep`: Temperature dependency
    - `alpha`: [K⁻¹] Temperature coefficient of resistance
    - `enforce_bounds`: Enforce bounds for the position of the wiper (0-1)

Analog Sensors

ModelingToolkitStandardLibrary.Electrical.CurrentSensor — Constant

CurrentSensor(; name)

Creates a circuit component that measures the current flowing through it. Analogous to an ideal ammeter.

States:

i(t): [A] Current through the sensor

Connectors:

p Positive pin

n Negative pin

ModelingToolkitStandardLibrary.Electrical.PotentialSensor — Constant

PotentialSensor(; name)

Creates a circuit component which measures the potential at a pin.

States:

phi(t): [V] The measured potential at this point

Connectors:

p Pin at which potential is to be measured

ModelingToolkitStandardLibrary.Electrical.VoltageSensor — Constant

VoltageSensor(; name)

Creates a circuit component that measures the voltage across it. Analogous to an ideal voltmeter.

States:

v(t): [V] The voltage difference from positive to negative pin p.v - n.v

Connectors:

p Positive pin

n Negative pin

ModelingToolkitStandardLibrary.Electrical.PowerSensor — Constant

PowerSensor(; name)

Combines a VoltageSensor and a CurrentSensor to measure the power being consumed by a circuit.

States:

power(t): [W] The power being consumed, given by the product of voltage and current

See VoltageSensor

See CurrentSensor

Connectors:

pc Corresponds to the p pin of the CurrentSensor

nc Corresponds to the n pin of the CurrentSensor

pv Corresponds to the p pin of the VoltageSensor

nv Corresponds to the n pin of the VoltageSensor

ModelingToolkitStandardLibrary.Electrical.MultiSensor — Constant

MultiSensor(; name)

Combines a VoltageSensor and a CurrentSensor.

States:

v(t): [V] The voltage across the VoltageSensor. Defaults to 1.0.

i(t): [A] The current across the CurrentSensor. Defaults to 1.0.

Connectors:

pc Corresponds to the p pin of the CurrentSensor

nc Corresponds to the n pin of the CurrentSensor

pv Corresponds to the p pin of the VoltageSensor

nv Corresponds to the n pin of the VoltageSensor

Analog Sources

ModelingToolkitStandardLibrary.Electrical.Voltage — Constant

Voltage(; name)

Acts as an ideal voltage source with no internal resistance.

States:

See OnePort

Connectors:

p Positive pin

n Negative pin

V RealInput Input for the voltage control signal, i.e. V ~ p.v - n.v

ModelingToolkitStandardLibrary.Electrical.Current — Constant

Current(; name)

Acts as an ideal current source with no internal resistance.

States:

See OnePort

Connectors:

p Positive pin

n Negative pin

I RealInput Input for the current control signal, i.e. `I ~ p.i

Digital Gates

Missing docstring.

Missing docstring for Not. Check Documenter's build log for details.

Missing docstring.

Missing docstring for And. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Nand. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Or. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Nor. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Xor. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Xnor. Check Documenter's build log for details.

Digital Components

Missing docstring.

Missing docstring for HalfAdder. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FullAdder. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MUX. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DEMUX. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Encoder. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Decoder. Check Documenter's build log for details.

Digital Sources

Missing docstring.

Missing docstring for PulseDiff. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Set. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Reset. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Pulse. Check Documenter's build log for details.