ModelingToolkitStandardLibrary: Electrical Components

ModelingToolkitStandardLibrary: Electrical Components

Index

Electrical Utilities

Analog Components

Analog Sensors

Analog Sources

Digital Gates

Digital Components

Digital Sources

Index

ModelingToolkitStandardLibrary.Electrical.Capacitor

ModelingToolkitStandardLibrary.Electrical.Conductor

ModelingToolkitStandardLibrary.Electrical.CurrentSensor

ModelingToolkitStandardLibrary.Electrical.DigitalPin

ModelingToolkitStandardLibrary.Electrical.Ground

ModelingToolkitStandardLibrary.Electrical.IdealOpAmp

ModelingToolkitStandardLibrary.Electrical.Inductor

ModelingToolkitStandardLibrary.Electrical.MultiSensor

ModelingToolkitStandardLibrary.Electrical.OnePort

ModelingToolkitStandardLibrary.Electrical.Pin

ModelingToolkitStandardLibrary.Electrical.PotentialSensor

ModelingToolkitStandardLibrary.Electrical.PowerSensor

ModelingToolkitStandardLibrary.Electrical.Resistor

ModelingToolkitStandardLibrary.Electrical.VoltageSensor

Electrical Utilities

ModelingToolkitStandardLibrary.Electrical.Pin — Function

Pin(; name)

A pin in an analog circuit.

States:

v(t): [V] The voltage at this pin

i(t): [A] The current passing through this pin

ModelingToolkitStandardLibrary.Electrical.OnePort — Function

OnePort(; name, v_start=0.0, i_start=0.0)

Component with two electrical pins p and n and current i flows from p to n.

States:

v(t): [V] The voltage across component p.v - n.v

i(t): [A] The current passing through positive pin

Parameters:

v_start: [V] Initial voltage across the component

i_start: [A] Initial current through the component

Connectors:

p Positive pin

n Negative pin

ModelingToolkitStandardLibrary.Electrical.DigitalPin — Function

DigitalPin(; name)

A pin in a digital circuit.

States:

v(t): [V] The voltage at this pin

i(t): [A] The current passing through this pin

val(t): The binary value of the pin at this point. A voltage from 0V to 0.8V is a binary value of 0.

A voltage in the range 2.0V to 5.0V is 1. Any other value is X.

Analog Components

ModelingToolkitStandardLibrary.Electrical.Ground — Function

Ground(; name)

Ground node with the potential of zero and connector g. Every circuit must have one ground node.

Connectors:

g

ModelingToolkitStandardLibrary.Electrical.Resistor — Function

Resistor(; name, R)

Creates an ideal Resistor following Ohm's Law.

States:

See OnePort

Connectors:

p Positive pin

n Negative pin

Parameters:

R: [Ohm] Resistance

ModelingToolkitStandardLibrary.Electrical.Conductor — Function

Conductor(;name, G)

Creates an ideal conductor.

States:

See OnePort

Connectors:

p Positive pin

n Negative pin

Parameters:

G: [S] Conductance

ModelingToolkitStandardLibrary.Electrical.Capacitor — Function

Capacitor(; name, C)

Creates an ideal capacitor.

States:

v(t): [V] The voltage across the capacitor, given by D(v) ~ p.i / C

Connectors:

p Positive pin

n Negative pin

Parameters:

C: [F] Capacitance

v_start: [V] Initial voltage of capacitor

ModelingToolkitStandardLibrary.Electrical.Inductor — Function

Inductor(; name, L)

Creates an ideal Inductor.

States:

See OnePort

Connectors:

p Positive pin

n Negative pin

Parameters:

L: [H] Inductance

i_start: [A] Initial current through inductor

ModelingToolkitStandardLibrary.Electrical.IdealOpAmp — Function

IdealOpAmp(; name)

Ideal operational amplifier (norator-nullator pair). The ideal OpAmp is a two-port. The left port is fixed to v1 = 0 and i1 = 0 (nullator). At the right port both any voltage v2 and any current i2 are possible (norator).

States:

See TwoPort

Connectors:

p1 Positive pin (left port)

p2 Positive pin (right port)

n1 Negative pin (left port)

n2 Negative pin (right port)

Analog Sensors

ModelingToolkitStandardLibrary.Electrical.CurrentSensor — Function

CurrentSensor(; name)

Creates a circuit component that measures the current flowing through it. Analogous to an ideal ammeter.

States:

i(t): [A] Current through the sensor

Connectors:

p Positive pin

n Negative pin

ModelingToolkitStandardLibrary.Electrical.PotentialSensor — Function

PotentialSensor(; name)

Creates a circuit component which measures the potential at a pin.

States:

phi(t): [V] The measured potential at this point

Connectors:

p Pin at which potential is to be measured

ModelingToolkitStandardLibrary.Electrical.VoltageSensor — Function

VoltageSensor(; name)

Creates a circuit component that measures the voltage across it. Analogous to an ideal voltmeter.

States:

v(t): [V] The voltage difference form positive to negative pin p.v - n.v

Connectors:

p Positive pin

n Negative pin

ModelingToolkitStandardLibrary.Electrical.PowerSensor — Function

PowerSensor(; name)

Combines a VoltageSensor and a CurrentSensor to measure the power being consumed by a circuit.

States:

power(t): [W] The power being consumed, given by the product of voltage and current.

See VoltageSensor

See CurrentSensor

Connectors:

pc Corresponds to the p pin of the CurrentSensor

nc Corresponds to the n pin of the CurrentSensor

pv Corresponds to the p pin of the VoltageSensor

nv Corresponds to the n pin of the VoltageSensor

ModelingToolkitStandardLibrary.Electrical.MultiSensor — Function

MultiSensor(; name)

Combines a VoltageSensor and a CurrentSensor.

States:

v(t): [V] The voltage across the VoltageSensor

i(t): [A] The current across the CurrentSensor

Connectors:

pc Corresponds to the p pin of the CurrentSensor

nc Corresponds to the n pin of the CurrentSensor

pv Corresponds to the p pin of the VoltageSensor

nv Corresponds to the n pin of the VoltageSensor

Analog Sources

Missing docstring.

Missing docstring for ConstantVoltage. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SineVoltage. Check Documenter's build log for details.

Missing docstring.

Missing docstring for StepVoltage. Check Documenter's build log for details.

Missing docstring.

Missing docstring for RampVoltage. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SquareVoltage. Check Documenter's build log for details.

Missing docstring.

Missing docstring for TriangularVoltage. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CosineVoltage. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ExpSineVoltage. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ConstantCurrent. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SineCurrent. Check Documenter's build log for details.

Missing docstring.

Missing docstring for StepCurrent. Check Documenter's build log for details.

Missing docstring.

Missing docstring for RampCurrent. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SquareCurrent. Check Documenter's build log for details.

Missing docstring.

Missing docstring for TriangularCurrent. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CosineCurrent. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ExpSineCurrent. Check Documenter's build log for details.

Digital Gates

Missing docstring.

Missing docstring for Not. Check Documenter's build log for details.

Missing docstring.

Missing docstring for And. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Nand. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Or. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Nor. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Xor. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Xnor. Check Documenter's build log for details.

Digital Components

Missing docstring.

Missing docstring for HalfAdder. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FullAdder. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MUX. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DEMUX. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Encoder. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Decoder. Check Documenter's build log for details.

Digital Sources

Missing docstring.

Missing docstring for PulseDiff. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Set. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Reset. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Pulse. Check Documenter's build log for details.