ModelingToolkit Standard Library: Mechanical Components

ModelingToolkit Standard Library: Mechanical Components

Index

Rotational Components

Rotational Utils

Rotational Core Components

Rotational Sources

Rotational Sensors

Index

ModelingToolkitStandardLibrary.Mechanical.Rotational.AngleSensor

ModelingToolkitStandardLibrary.Mechanical.Rotational.Damper

ModelingToolkitStandardLibrary.Mechanical.Rotational.Fixed

ModelingToolkitStandardLibrary.Mechanical.Rotational.Flange

ModelingToolkitStandardLibrary.Mechanical.Rotational.IdealGear

ModelingToolkitStandardLibrary.Mechanical.Rotational.Inertia

ModelingToolkitStandardLibrary.Mechanical.Rotational.PartialCompliant

ModelingToolkitStandardLibrary.Mechanical.Rotational.PartialCompliantWithRelativeStates

ModelingToolkitStandardLibrary.Mechanical.Rotational.PartialElementaryOneFlangeAndSupport2

ModelingToolkitStandardLibrary.Mechanical.Rotational.PartialElementaryTwoFlangesAndSupport2

ModelingToolkitStandardLibrary.Mechanical.Rotational.RelSpeedSensor

ModelingToolkitStandardLibrary.Mechanical.Rotational.RotationalFriction

ModelingToolkitStandardLibrary.Mechanical.Rotational.Speed

ModelingToolkitStandardLibrary.Mechanical.Rotational.SpeedSensor

ModelingToolkitStandardLibrary.Mechanical.Rotational.Spring

ModelingToolkitStandardLibrary.Mechanical.Rotational.SpringDamper

ModelingToolkitStandardLibrary.Mechanical.Rotational.Support

ModelingToolkitStandardLibrary.Mechanical.Rotational.Torque

ModelingToolkitStandardLibrary.Mechanical.Rotational.TorqueSensor

Rotational Components

Rotational Utils

ModelingToolkitStandardLibrary.Mechanical.Rotational.Flange — Function

Support(;name)

1-dim. rotational flange of a shaft.

States:

phi(t): [rad] Absolute rotation angle of flange

tau(t): [N.m] Cut torque in the flange

ModelingToolkitStandardLibrary.Mechanical.Rotational.Support — Function

Support(;name)

Support/housing of a 1-dim. rotational shaft

States:

phi(t): [rad] Absolute rotation angle of the support/housing

tau(t): [N.m] Cut torque in the support/housing

ModelingToolkitStandardLibrary.Mechanical.Rotational.PartialCompliantWithRelativeStates — Function

PartialCompliantWithRelativeStates(;name, phi_rel_start=0.0, tau_start=0.0)

Partial model for the compliant connection of two rotational 1-dim. shaft flanges where the relative angle and speed are used as preferred states

States:

phi_rel(t): [rad] Relative rotation angle (= flangeb.phi - flangea.phi)

w_rel(t): [rad/s] Relative angular velocity (= D(phi_rel))

a_rel(t): [rad/s²] Relative angular acceleration (= D(w_rel))

tau(t): [N.m] Torque between flanges (= flange_b.tau)

Connectors:

flange_a Flange

flange_b Flange

Parameters:

phi_rel_start: [rad] Initial relative rotation angle

w_rel_start: [rad/s] Initial relative angular velocity (= D(phi_rel))

a_rel_start: [rad/s²] Initial relative angular acceleration (= D(w_rel))

tau_start: [N.m] Initial torque between flanges

ModelingToolkitStandardLibrary.Mechanical.Rotational.PartialElementaryOneFlangeAndSupport2 — Function

PartialElementaryOneFlangeAndSupport2(;name, use_support=false)

Partial model for a component with one rotational 1-dim. shaft flange and a support used for textual modeling, i.e., for elementary models

States:

phi_support(t): [rad] Absolute angle of support flange"

Connectors:

flange Flange

Parameters:

use_support: If support flange enabled, otherwise implicitly grounded

ModelingToolkitStandardLibrary.Mechanical.Rotational.PartialElementaryTwoFlangesAndSupport2 — Function

PartialElementaryTwoFlangesAndSupport2(;name, use_support=false)

Partial model for a component with two rotational 1-dim. shaft flanges and a support used for textual modeling, i.e., for elementary models

States:

phi_support(t): [rad] Absolute angle of support flange

Connectors:

flange_a Flange

flange_b Flange

support Support if use_support == true

Parameters:

use_support: If support flange enabled, otherwise implicitly grounded

ModelingToolkitStandardLibrary.Mechanical.Rotational.PartialCompliant — Function

PartialCompliant(;name, phi_rel_start=0.0, tau_start=0.0)

Partial model for the compliant connection of two rotational 1-dim. shaft flanges.

States:

phi_rel(t): [rad] Relative rotation angle (flange_b.phi - flange_a.phi)

tau(t): [N.m] Torque between flanges (flange_b.tau)

Connectors:

flange_a Flange

flange_b Flange

Parameters:

phi_rel_start: [rad] Initial relative rotation angle

tau_start: [N.m] Initial torque between flanges

Rotational Core Components

ModelingToolkitStandardLibrary.Mechanical.Rotational.Fixed — Function

Fixed(;name, phi0=0.0)

Flange fixed in housing at a given angle.

Connectors:

flange Flange

Parameters:

phi0: [rad] Fixed offset angle of housing

ModelingToolkitStandardLibrary.Mechanical.Rotational.Inertia — Function

Inertia(;name, J, phi_start=0.0, w_start=0.0, a_start=0.0)

1D-rotational component with inertia.

States:

phi: [rad] Absolute rotation angle of component

w: [rad/s] Absolute angular velocity of component (= D(phi))

a: [rad/s²] Absolute angular acceleration of component (= D(w))

Connectors:

flange_a Flange Left flange

flange_b Flange Right flange

Parameters:

J: [kg·m²] Moment of inertia

phi_start: [rad] Initial value of absolute rotation angle of component

w_start: [rad/s] Initial value of absolute angular velocity of component

a_start: [rad/s²] Initial value of absolute angular acceleration of component

ModelingToolkitStandardLibrary.Mechanical.Rotational.Spring — Function

Spring(;name, c, phi_rel0=0.0)

Linear 1D rotational spring

States:

phi_rel(t): [rad] Relative rotation angle (flange_b.phi - flange_a.phi)

tau(t): [N.m] Torque between flanges (flange_b.tau)

Connectors:

flange_a Flange

flange_b Flange

Parameters:

c: [N.m/rad] Spring constant

phi_rel0: [rad] Unstretched spring angle

ModelingToolkitStandardLibrary.Mechanical.Rotational.Damper — Function

Damper(;name, d)

Linear 1D rotational damper

States:

phi_rel(t): [rad] Relative rotation angle (= flangeb.phi - flangea.phi)

w_rel(t): [rad/s] Relative angular velocity (= D(phi_rel))

a_rel(t): [rad/s²] Relative angular acceleration (= D(w_rel))

tau(t): [N.m] Torque between flanges (= flange_b.tau)

Connectors:

flange_a Flange

flange_b Flange

Parameters:

d: [N.m.s/rad] Damping constant

ModelingToolkitStandardLibrary.Mechanical.Rotational.SpringDamper — Function

SpringDamper(;name, d)

Linear 1D rotational spring and damper

States:

phi_rel(t): [rad] Relative rotation angle (= flangeb.phi - flangea.phi)

w_rel(t): [rad/s] Relative angular velocity (= D(phi_rel))

a_rel(t): [rad/s²] Relative angular acceleration (= D(w_rel))

tau(t): [N.m] Torque between flanges (= flange_b.tau)

Connectors:

flange_a Flange

flange_b Flange

Parameters:

d: [N.m.s/rad] Damping constant

c: [N.m/rad] Spring constant

ModelingToolkitStandardLibrary.Mechanical.Rotational.IdealGear — Function

IdealGear(;name, ratio, use_support=false)

Ideal gear without inertia.

This element characterizes any type of gear box which is fixed in the ground and which has one driving shaft and one driven shaft.

States:

phi_a(t): [rad] Relative angle between shaft a and the support

phi_b(t): [rad] Relative angle between shaft b and the support

Connectors:

flange_a Flange

flange_b Flange

support Support if use_support == true

Parameters:

ratio: Transmission ratio (flangea.phi/flangeb.phi)

use_support: If support flange enabled, otherwise implicitly grounded

ModelingToolkitStandardLibrary.Mechanical.Rotational.RotationalFriction — Function

RotationalFriction(;name, f, tau_c, w_brk, tau_brk)

Models rotational friction with Stribeck effect, Coulomb friction and viscous friction between the two flanges. The friction torque is a function of the relative angular velocity between flangea and flangeb.

Friction model: "Armstrong, B. and C.C. de Wit, Friction Modeling and Compensation, The Control Handbook, CRC Press, 1995."

States:

phi_rel(t): [rad] Relative rotation angle (= flangeb.phi - flangea.phi)

w_rel(t): [rad/s] Relative angular velocity (= D(phi_rel))

a_rel(t): [rad/s²] Relative angular acceleration (= D(w_rel))

tau(t): [N.m] Torque between flanges (= flange_b.tau)

Connectors:

flange_a Flange

flange_b Flange

Parameters:

f: [N⋅m/(rad/s)] Viscous friction coefficient

tau_c: [N⋅m] Coulomb friction torque

w_brk: [rad/s] Breakaway friction velocity

tau_brk: [N⋅m] Breakaway friction torque

Rotational Sources

ModelingToolkitStandardLibrary.Mechanical.Rotational.Torque — Function

Torque(; name, use_support=false)

Input signal acting as external torque on a flange

States:

phi_support(t): [rad] Absolute angle of support flange

Connectors:

flange Flange

tau RealInput Accelerating torque acting at flange -flange.tau

Parameters:

use_support

ModelingToolkitStandardLibrary.Mechanical.Rotational.Speed — Function

Speed(; name, use_support=false, exact=false, f_crit=50)

Forced movement of a flange according to a reference angular velocity signal

States:

phi_support(t): [rad] Absolute angle of support flange"

Connectors:

flange Flange

w_ref RealInput Reference angular velocity of flange with respect to support as input signal needs to be continuously differential

Parameters:

use_support: If support flange enabled, otherwise implicitly grounded

exact: true/false exact treatment/filtering the input signal

tau_filt: [rad/s] if exact=false, Time constant of low-pass filter to filter input signal

Rotational Sensors

ModelingToolkitStandardLibrary.Mechanical.Rotational.AngleSensor — Function

AngleSensor(;name)

Ideal sensor to measure the absolute flange angle

Connectors:

flange: Flange Flange of shaft from which sensor information shall be measured

phi: RealOutput Absolute angle of flange

ModelingToolkitStandardLibrary.Mechanical.Rotational.SpeedSensor — Function

SpeedSensor(;name)

Ideal sensor to measure the absolute flange angular velocity

Connectors:

flange: Flange Flange of shaft from which sensor information shall be measured

w: RealOutput Absolute angular velocity of flange

ModelingToolkitStandardLibrary.Mechanical.Rotational.TorqueSensor — Function

TorqueSensor(;name)

Ideal sensor to measure the torque between two flanges (= flange_a.tau)

Connectors:

flange_a: Flange Left flange of shaft

flange_b: Flange Left flange of shaft

tau: RealOutput Torque in flange flangea and flangeb (tau = flange_a.tau = -flange_b.tau)

ModelingToolkitStandardLibrary.Mechanical.Rotational.RelSpeedSensor — Function

RelSpeedSensor(;name)

Ideal sensor to measure the relative angular velocity

Connectors:

flange_a: Flange Flange of shaft from which sensor information shall be measured

flange_b: Flange Flange of shaft from which sensor information shall be measured

w: RealOutput Absolute angular velocity of flange